Moved internal config + began route expansion
